Refactor run module into modular structure
- Split monolithic run.rs into focused modules:
  - run/environment.rs: TestEnvironment and file operations
  - run/params.rs: RunParams configuration and conditions
  - run/commands.rs: Built-in command implementations
  - run/execution.rs: Command dispatch and execution logic
  - run/mod.rs: Public API coordination

- Improves maintainability and makes codebase easier to navigate
- All 37 tests still pass
- No functionality changes, pure refactoring
